move auth to it's own package
