+++ auth/consts.go
