+++ auth/ed25519.go
