+++ examples/morpheusvm/registry/registry.go
@@ -6,8 +6,8 @@ package registry
 import (
 	"github.com/ava-labs/avalanchego/utils/wrappers"
 
+	"github.com/ava-labs/hypersdk/auth"
 	"github.com/ava-labs/hypersdk/chain"
-	"github.com/ava-labs/hypersdk/chain/auth"
 	"github.com/ava-labs/hypersdk/codec"
 	"github.com/ava-labs/hypersdk/examples/morpheusvm/actions"
 	"github.com/ava-labs/hypersdk/examples/morpheusvm/consts"
