+++ examples/morpheusvm/tests/e2e/e2e_test.go
@@ -18,8 +18,8 @@ import (
 	"github.com/fatih/color"
 	"github.com/stretchr/testify/require"
 
+	"github.com/ava-labs/hypersdk/auth"
 	"github.com/ava-labs/hypersdk/chain"
-	"github.com/ava-labs/hypersdk/chain/auth"
 	"github.com/ava-labs/hypersdk/codec"
 	"github.com/ava-labs/hypersdk/crypto/ed25519"
 	"github.com/ava-labs/hypersdk/examples/morpheusvm/actions"
