+++ examples/tokenvm/cmd/token-faucet/manager/manager.go
@@ -16,8 +16,8 @@ import (
 	"github.com/ava-labs/avalanchego/utils/timer"
 	"go.uber.org/zap"
 
+	"github.com/ava-labs/hypersdk/auth"
 	"github.com/ava-labs/hypersdk/chain"
-	"github.com/ava-labs/hypersdk/chain/auth"
 	"github.com/ava-labs/hypersdk/codec"
 	"github.com/ava-labs/hypersdk/examples/tokenvm/actions"
 	"github.com/ava-labs/hypersdk/examples/tokenvm/challenge"
